refactor: extract models and enums to _models.py
Move enums (Outputs, RecordTypes, SessionMode) and dataclasses
(SystemInfo, ProcessStats, LogPaths, Averages, Sample) to _models.py.
Create _utils.py with assert_num helper.

Part of #372
