+++ src/con_duct/_utils.py
@@ -1,0 +1,8 @@
+"""Utility functions for con-duct."""
+
+from typing import Any
+
+
+def assert_num(*values: Any) -> None:
+    for value in values:
+        assert isinstance(value, (float, int))
